Do not log access tokens and unredacted MAA tokens even in debug logging
This is highly surprising and can easily end up leaking it if the log is made
available in unprotected places.

For AAD tokens, simply do not log them, since if a token is generated
successfully, it is not likely to be useful for further debugging to see the
token.

For MAA tokens, this PR logs the information part but redacts the signature.