Do not log access tokens and keys in debug mode (#227)
* Do not log access tokens and unredacted MAA tokens even in debug logging

This is highly surprising and can easily end up leaking it if the log is made
available in unprotected places.

For AAD tokens, simply do not log them, since if a token is generated
successfully, it is not likely to be useful for further debugging to see the
token.

For MAA tokens, this PR logs the information part but redacts the signature.

* Do not log released keys in debug mode

Like the previous commit, this also prevents accidental exposure of the keys,
especially to the (untrusted) host.

* Add Test_RedactMAAToken

* remotefs: log some data from input json